+++ honeycomb-kernels/src/grisubal/routines/compute_intersecs.rs
