+++ honeycomb-kernels/src/grisubal/routines/process_intersecs_data.rs
